Autostart eclimd
Start eclimd from within Emacs whenever eclim-mode is enabled or
global-eclim-mode requires eclimd to determine if eclim-mode should be enabled
in a buffer. When eclimd was started outside of Emacs it is detected and eclimd
does not get started a second time. Autostarting can be disabled by setting
eclimd-autostart to nil, eclimd-autostart-with-default-workspace controls
whether eclimd asks for a workspace when it gets autostarted.