Travis build passes again
To use a different version of Emacs with Cask, the path to the Emacs
executable needs to be specified in the EMACS environment variable
now. These changes were made in the Travis CI script.

Also, instead of downloading the shell script
https://gist.github.com/rejeep/ebcd57c3af83b049833b/raw, the commands
are now in the Travic CI script.

Allow Emacs versoin emacs-git-snapshot-travis to fail for now.